backInventario/backend/requirements.txt
# Dependencias del Backend - Sistema de Heladería

# Framework web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Validación de datos
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS y seguridad
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Variables de entorno
python-dotenv==1.0.0

# Testing (opcional)
pytest==7.4.3
httpx==0.25.2
